chore: Add custom hook for handling directions.
