+++ packages/map-template/src/atoms/hasFoundRouteState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const hasFoundRouteState = atom({
+    key: 'hasFoundRoute',
+    default: true
+});
+
+export default hasFoundRouteState;
