+++ packages/map-template/src/components/LocationDetails/LocationDetails.scss
@@ -100,7 +100,6 @@
 
         &:hover {
             filter: brightness(90%);
-            cursor: pointer;
         }
     }
