+++ packages/map-template/src/components/Wayfinding/Wayfinding.jsx
@@ -27,10 +27,10 @@ import { ReactComponent as CompassArrow } from '../../assets/compass-arrow.svg';
 import { travelModes } from "../../constants/travelModes";
 import Dropdown from "../WebComponentWrappers/Dropdown/Dropdown";
 import primaryColorState from "../../atoms/primaryColorState";
-import directionsResponseState from "../../atoms/directionsResponseState";
 import addMapboxPlaceGeometry from "../Map/MapboxMap/MapboxPlacesHandler";
 import mapboxAccessTokenState from "../../atoms/mapboxAccessTokenState";
-import getLocationPoint from "../../helpers/GetLocationPoint";
+import useDirectionsInfo from "../../hooks/useDirectionsInfo";
+import hasFoundRouteState from "../../atoms/hasFoundRouteState";
 
 const searchFieldIdentifiers = {
     TO: 'TO',
@@ -61,7 +61,6 @@ function Wayfinding({ onStartDirections, onBack, directionsToLocation, direction
     const toFieldRef = useRef();
     const fromFieldRef = useRef();
 
-    const [, setDirectionsResponse] = useRecoilState(directionsResponseState);
     const directionsService = useRecoilValue(directionsServiceState);
     const userPosition = useRecoilValue(userPositionState);
     const currentLocation = useRecoilValue(currentLocationState);
@@ -71,7 +70,7 @@ function Wayfinding({ onStartDirections, onBack, directionsToLocation, direction
     const [activeSearchField, setActiveSearchField] = useState();
 
     /** Indicate if a route has been found */
-    const [hasFoundRoute, setHasFoundRoute] = useState(true);
+    const [, setHasFoundRoute] = useRecoilState(hasFoundRouteState);
 
     /** Indicate if search results have been found */
     const [hasSearchResults, setHasSearchResults] = useState(true);
@@ -84,9 +83,6 @@ function Wayfinding({ onStartDirections, onBack, directionsToLocation, direction
 
     const [destinationLocation, setDestinationLocation] = useState();
     const [originLocation, setOriginLocation] = useState();
-
-    const [totalDistance, setTotalDistance] = useState();
-    const [totalTime, setTotalTime] = useState();
 
     const [accessibilityOn, setAccessibilityOn] = useState(false);
 
@@ -101,6 +97,7 @@ function Wayfinding({ onStartDirections, onBack, directionsToLocation, direction
 
     const mapboxAccessToken = useRecoilValue(mapboxAccessTokenState);
 
+    const [totalDistance, totalTime, hasFoundRoute] = useDirectionsInfo(originLocation, destinationLocation, directionsService, travelMode, accessibilityOn)
 
     /**
      * Decorates location with data that is required for wayfinding to work.
@@ -348,43 +345,6 @@ function Wayfinding({ onStartDirections, onBack, directionsToLocation, direction
             }
         }
     }, [isActive, directionsToLocation, directionsFromLocation]);
-
-    /*
-     * When both origin location and destination location are selected, and have geometry, call the MapsIndoors SDK
-     * to get information about the route.
-     */
-    useEffect(() => {
-        if (originLocation?.geometry && destinationLocation?.geometry) {
-            directionsService.getRoute({
-                origin: getLocationPoint(originLocation),
-                destination: getLocationPoint(destinationLocation),
-                travelMode: travelMode,
-                avoidStairs: accessibilityOn
-            }).then(directionsResult => {
-                if (directionsResult && directionsResult.legs) {
-                    setHasFoundRoute(true);
-                    // Calculate total distance and time
-                    const totalDistance = directionsResult.legs.reduce((accumulator, current) => accumulator + current.distance.value, 0);
-                    const totalTime = directionsResult.legs.reduce((accumulator, current) => accumulator + current.duration.value, 0);
-
-                    setTotalDistance(totalDistance);
-                    setTotalTime(totalTime);
-
-                    setDirectionsResponse({
-                        originLocation,
-                        destinationLocation,
-                        totalDistance,
-                        totalTime,
-                        directionsResult
-                    });
-                } else {
-                    setHasFoundRoute(false);
-                }
-            }, () => {
-                setHasFoundRoute(false);
-            });
-        }
-    }, [originLocation, destinationLocation, directionsService, accessibilityOn, travelMode]);
 
     /*
      * React on changes on the selected map type.
